+++ frontend/src/features/login/Screens/SignUp/SignUp0202.js
@@ -1,15 +1,25 @@
 import React, { useState } from "react";
 import { StyleSheet, TouchableOpacity, View, Text } from "react-native";
-import { theme } from "../../../infrastructure/theme";
+import { theme } from "../../../../infrastructure/theme";
 import { SvgXml } from "react-native-svg";
 import { TextInput } from "react-native-gesture-handler";
-import backButtonWhite from "../../../../assets/Buttons/backButtonWhite";
-import { LoginBg } from "../Component/LoginBg";
-import { LoginButton } from "../Component/LoginButton";
+import backButtonWhite from "../../../../../assets/Buttons/backButtonWhite";
+import { LoginBg } from "../../Component/LoginBg";
+import { LoginButton } from "../../Component/LoginButton";
 
-export const SignUp0205 = ({ navigation }) => {
+export const SignUp0202 = ({ navigation, route }) => {
+  const { userInfo, setUserInfo } = route.params;
+
+  const handleName = (e) => {
+    setUserInfo((state) => {
+      return { ...state, name: e };
+    });
+  };
+
   const nextButton = async () => {
-    navigation.navigate("MapScreen");
+    // 이름 입력받음
+    // SignUp0203로 넘어감
+    navigation.navigate("SignUp0203", { userInfo, setUserInfo });
   };
 
   return (
@@ -27,16 +37,23 @@ export const SignUp0205 = ({ navigation }) => {
 
         {/* Main 안내 문구 */}
         <View style={styles.container2}>
-          <Text style={styles.mainText}>
-            두근두근, 드롭바이와{`\n`}여행을 떠나보실까요?
-          </Text>
+          <Text style={styles.mainText}>이름이 뭐에요?</Text>
         </View>
 
         {/* 유저 입력창 */}
         <View style={styles.container3}>
+          <View style={styles.inputBox}>
+            <TextInput
+              style={styles.input}
+              placeholderTextColor="#02B5AA"
+              placeholder=""
+              onChangeText={(name) => handleName(name)}
+              value={userInfo.name}
+            ></TextInput>
+          </View>
           <LoginButton
-            style={{ marginTop: 24 }}
-            value="네!"
+            style={{ marginTop: 100 }}
+            value="다 음"
             onPress={nextButton}
             width={300}
             height={43}
@@ -57,12 +74,12 @@ const styles = StyleSheet.create({
     alignItems: "center",
   },
   container2: {
-    flex: 2,
-    justifyContent: "flex-end",
+    flex: 1,
+    justifyContent: "center",
     alignItems: "center",
   },
   container3: {
-    flex: 3,
+    flex: 4,
     justifyContent: "flex-start",
     alignItems: "center",
   },
@@ -71,13 +88,15 @@ const styles = StyleSheet.create({
     flex: 0,
     alignItems: "center",
     justifyContent: "center",
+
     color: "white",
     fontFamily: theme.fonts.bold,
     fontWeight: "700",
-    fontSize: 24,
+    fontSize: 20,
   },
 
   input: {
+    width: 100,
     justifyContent: "center",
     alignItems: "center",
     fontSize: 24,
